# internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ash-repwiki_ash_internal_ci.DiagnosisResponse:
    properties:
      adopted:
        type: boolean
      confidence:
        type: number
      connectionId:
        type: string
      createdAt:
        type: string
      decidedAt:
        type: string
      decidedBy:
        type: string
      decisionReason:
        type: string
      decisionStatus:
        type: string
      evidenceRefs:
        items:
          type: string
        type: array
      fixSuggestions:
        items:
          type: string
        type: array
      id:
        type: string
      jobId:
        type: string
      logDigest:
        type: string
      rootCause:
        type: string
      runId:
        type: string
      spaceId:
        type: string
      status:
        type: string
    type: object
...
  internal_api.APIError:
    properties:
      code:
        example: INVALID_REQUEST
        type: string
      message:
        example: validation failed
        type: string
    type: object
  internal_api.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_api.APIError'
    type: object
...
  internal_api.AuthGatewaySession:
    properties:
      did:
        type: string
      exp:
        type: integer
      scope:
        items:
          type: string
        type: array
      sid:
        type: string
      spaceId:
        type: string
      status:
        type: string
      typ:
        type: string
    type: object
  internal_api.AuthMeResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
      session:
        $ref: '#/definitions/internal_api.AuthGatewaySession'
      space:
        $ref: '#/definitions/internal_api.AuthSpace'
      user:
        $ref: '#/definitions/internal_api.AuthUser'
    type: object
  internal_api.AuthSessionResponse:
    properties:
      refreshToken:
        type: string
      session:
        $ref: '#/definitions/internal_api.AuthGatewaySession'
      space:
        $ref: '#/definitions/internal_api.AuthSpace'
      token:
        type: string
      user:
        $ref: '#/definitions/internal_api.AuthUser'
    type: object
  internal_api.AuthSpace:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  internal_api.AuthUser:
    properties:
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
...
  internal_api.PasswordChangeResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  internal_api.changePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
...
  internal_api.githubWebhookResponse:
    properties:
      ashRunId:
        type: string
      ashTraceId:
        type: string
      ciJobId:
        type: string
      ciRunId:
        type: string
      conclusion:
        type: string
      connectionId:
        type: string
      diagnosis:
        $ref: '#/definitions/github_com_ash-repwiki_ash_internal_ci.DiagnosisResponse'
      duplicate:
        type: boolean
      executionError:
        type: string
      ignored:
        type: boolean
      planId:
        type: string
      providerRunId:
        type: string
      reason:
        type: string
      shouldStartRun:
        type: boolean
      spaceId:
        type: string
      workflow:
        type: string
    type: object
...
  internal_api.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      spaceId:
        type: string
    required:
    - email
    - password
    type: object
...
  internal_api.refreshAuthSessionRequest:
    properties:
      refreshToken:
        type: string
      ttlSeconds:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: ASH
  description: Delivery-oriented coding robot worker API (M0). See docs/appendices/G-OpenAPI-端点清单(M0).md.
  license:
    name: Proprietary
  title: ASH Worker API
  version: "0.1"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AuthSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Login with local credentials
      tags:
      - auth
  /api/v1/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AuthMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Get current authenticated identity
      tags:
      - auth
  /api/v1/auth/oidc/callback:
    get:
      description: Exchanges code for id_token claims (HS256 or RS256/JWKS), links/JIT
        user, returns ASH JWT session.
      parameters:
      - description: authorization code
        in: query
        name: code
        required: true
        type: string
      - description: csrf state
        in: query
        name: state
        required: true
        type: string
      - description: preferred space id
        in: query
        name: spaceId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AuthSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: OIDC authorization-code callback
      tags:
      - auth
  /api/v1/auth/oidc/login:
    get:
      description: Redirects to the configured IdP when ASH_OIDC_ENABLED=1. Pass ui=1
        for console hash redirect on callback.
      parameters:
      - description: set to 1 for /ui/login hash redirect after callback
        in: query
        name: ui
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: redirect to IdP
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Start OIDC authorization-code login
      tags:
      - auth
  /api/v1/auth/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: password change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.PasswordChangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Change current user's password
      tags:
      - auth
...
  /api/v1/auth/sessions/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refreshToken and optional access ttl
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_api.refreshAuthSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AuthSessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Rotate access+refresh JWTs for an active auth session (same sid)
      tags:
      - auth
...
  /api/v1/webhooks/github:
    post:
      consumes:
      - application/json
      description: 'Public path: verifies X-Hub-Signature-256, upserts CI run/job,
        diagnoses failures, creates a GoalPlan draft (Quest); autoRun=1 auto-approves
        into a hotfix Run.'
      parameters:
      - description: repo connection id
        in: query
        name: connectionId
        required: true
        type: string
      - description: auto-approve GoalPlan into hotfix run on failure
        in: query
        name: autoRun
        type: boolean
      - default: .
        description: repo root for GoalPlan / autoRun
        in: query
        name: repoRoot
        type: string
      - description: sha256 HMAC of raw body
        in: header
        name: X-Hub-Signature-256
        required: true
        type: string
      - description: delivery id for idempotency
        in: header
        name: X-GitHub-Delivery
        type: string
      - default: workflow_run
        description: event name
        in: header
        name: X-GitHub-Event
        type: string
      - description: GitHub workflow_run payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.githubWebhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIErrorResponse'
      summary: Ingest GitHub Actions webhook (HMAC)
      tags:
      - webhooks
...
swagger: "2.0"
...